requirements.txt
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.5
seaborn==0.13.2
scikit-learn==1.3.2
xgboost==2.1.4
python-dotenv==1.0.1
dill==0.4.0
google-cloud-bigquery==3.30.0
google-api-python-client
db-dtypes==1.4.2
imblearn==0.0
gunicorn==23.0.0
streamlit==1.40.1
ipykernel==6.29.5
tldextract==5.1.2
jupyter==1.1.1
plotly==6.0.1
python-whois== 0.9.5
requests==2.32.3
beautifulsoup4
tld==0.12.6
ipwhois==1.3.0
aslookup==1.3.0
dnspython==2.6.1
db-dtypes
-e .